requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
keras==2.15.0
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
newsapi-python==0.2.7
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
statsmodels==0.14.0
pymongo==4.6.0
python-multipart==0.0.6
pydantic==2.5.0
motor==3.3.2
ta==0.11.0
python-dotenv==1.0.0
schedule==1.2.1